move multilabel classification out of text folder
