+++ mteb/abstasks/__init__.py
@@ -5,10 +5,10 @@ from .any_sts import AbsTaskAnySTS
 from .any_zeroshot_classification import AbsTaskAnyZeroShotClassification
 from .clustering_fast import AbsTaskClusteringFast
 from .image.image_text_pair_classification import AbsTaskImageTextPairClassification
+from .multilabel_classification import AbsTaskMultilabelClassification
 from .regression import AbsTaskRegression
 from .retrieval import AbsTaskRetrieval
 from .text.bitext_mining import AbsTaskBitextMining
-from .text.multilabel_classification import AbsTaskMultilabelClassification
 from .text.pair_classification import AbsTaskPairClassification
 from .text.reranking import AbsTaskReranking
 from .text.summarization import AbsTaskSummarization
